solution 이름 변경
- 기존 ObserverPattern -> Redux로
- 새로 추가된 ObserverPattern2 -> ObserverPattern으로 이름 변경
