+++ ObserverPattern/Manager.cpp
