+++ ObserverPattern/Manager.h
