+++ ObserverPattern/Observer.cpp
