--- ObserverPattern2/Debug/ObserverPattern2.tlog/ObserverPattern2.lastbuildstate
@@ -1,2 +1,0 @@
-#TargetFrameworkVersion=v4.0:PlatformToolSet=v142:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=10.0
-Debug|Win32|C:\Users\ohhh1\Desktop\Coding\DesignPattern\ObserverPattern2\|
