--- ObserverPattern2/ObserverPattern2.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29613.14
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ObserverPattern2", "ObserverPattern2.vcxproj", "{E36A6929-A18F-4589-89F9-C1CBF3D22AF9}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E36A6929-A18F-4589-89F9-C1CBF3D22AF9}.Debug|x64.ActiveCfg = Debug|x64
-		{E36A6929-A18F-4589-89F9-C1CBF3D22AF9}.Debug|x64.Build.0 = Debug|x64
-		{E36A6929-A18F-4589-89F9-C1CBF3D22AF9}.Debug|x86.ActiveCfg = Debug|Win32
-		{E36A6929-A18F-4589-89F9-C1CBF3D22AF9}.Debug|x86.Build.0 = Debug|Win32
-		{E36A6929-A18F-4589-89F9-C1CBF3D22AF9}.Release|x64.ActiveCfg = Release|x64
-		{E36A6929-A18F-4589-89F9-C1CBF3D22AF9}.Release|x64.Build.0 = Release|x64
-		{E36A6929-A18F-4589-89F9-C1CBF3D22AF9}.Release|x86.ActiveCfg = Release|Win32
-		{E36A6929-A18F-4589-89F9-C1CBF3D22AF9}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {F1BCF178-1940-4A5A-ACC1-B40317152E46}
-	EndGlobalSection
-EndGlobal
