--- ObserverPattern2/ObserverPattern2.vcxproj.filters
@@ -1,39 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <Filter Include="Source">
-      <UniqueIdentifier>{cce581c4-3050-43dc-a80a-8696e7b74342}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Source\0_main">
-      <UniqueIdentifier>{c2b4e7b5-c6ff-4dfe-80c8-1cb875d52d5a}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Source\1_Manager">
-      <UniqueIdentifier>{7e7b68ad-056d-4fc0-b534-8f926e93eee5}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Source\2_Observer">
-      <UniqueIdentifier>{df449262-39f7-4376-976f-2b1f3be607a0}</UniqueIdentifier>
-    </Filter>
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="Main.cpp">
-      <Filter>Source\0_main</Filter>
-    </ClCompile>
-    <ClCompile Include="Manager.cpp">
-      <Filter>Source\1_Manager</Filter>
-    </ClCompile>
-    <ClCompile Include="Observer.cpp">
-      <Filter>Source\2_Observer</Filter>
-    </ClCompile>
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="Project.h">
-      <Filter>Source\0_main</Filter>
-    </ClInclude>
-    <ClInclude Include="Manager.h">
-      <Filter>Source\1_Manager</Filter>
-    </ClInclude>
-    <ClInclude Include="Observer.h">
-      <Filter>Source\2_Observer</Filter>
-    </ClInclude>
-  </ItemGroup>
-</Project>
