+++ Redux+Observer/CManager.cpp
