+++ Redux+Observer/CManager.h
