+++ Redux+Observer/CObserver.cpp
