+++ Redux+Observer/CSingleton.cpp
