+++ Redux+Observer/Object.cpp
