+++ Redux+Observer/Object.h
