+++ Redux+Observer/ObserverPattern.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29001.49
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ObserverPattern", "ObserverPattern.vcxproj", "{E2D6097D-6A07-47AB-A930-222E980D72E3}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E2D6097D-6A07-47AB-A930-222E980D72E3}.Debug|x64.ActiveCfg = Debug|x64
+		{E2D6097D-6A07-47AB-A930-222E980D72E3}.Debug|x64.Build.0 = Debug|x64
+		{E2D6097D-6A07-47AB-A930-222E980D72E3}.Debug|x86.ActiveCfg = Debug|Win32
+		{E2D6097D-6A07-47AB-A930-222E980D72E3}.Debug|x86.Build.0 = Debug|Win32
+		{E2D6097D-6A07-47AB-A930-222E980D72E3}.Release|x64.ActiveCfg = Release|x64
+		{E2D6097D-6A07-47AB-A930-222E980D72E3}.Release|x64.Build.0 = Release|x64
+		{E2D6097D-6A07-47AB-A930-222E980D72E3}.Release|x86.ActiveCfg = Release|Win32
+		{E2D6097D-6A07-47AB-A930-222E980D72E3}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {1D4F0A6B-C97B-41AF-B2FC-9739A9A1AF4A}
+	EndGlobalSection
+EndGlobal
