+++ Redux+Observer/ObserverPattern.vcxproj.user
