+++ Redux+Observer/Singleton.h
